+++ server/src/main/java/org/elasticsearch/script/field/vectors/ByteKnnDenseVector.java
@@ -51,7 +51,7 @@ public class ByteKnnDenseVector implements DenseVector {
     }
 
     @Override
-    public int dotProduct(byte[] queryVector) {
+    public double dotProduct(byte[] queryVector) {
         return ESVectorUtil.dotProduct(queryVector, docVector);
     }
